Fix computer property already defined warning
Fixes 'The computed property "PageNames" is already
defined in data.'on the lesson summary page in Coach.

This happens because 'LessonSummaryPage' has
'PageNames' in its 'data' while at the same time importing
'commonCoach' mixin that contains 'PageNames' as computed.

Also adds 'this.' before 'PageNames' so there's no need
for the global import.