Fix unintended lack of space
around the description input.

This was caused by -1em margin left
being applied to all textboxes instead
of only the title box.

This fix ensures that this style is
applied to the title box only, and only
on large screens. On smaller screens it
would cause issues with lack of space
between the input and icon.