LessonResourceSelection: Add margin to bottom controls when isAppContext && isTouchDevice
In Quiz this is not necessary because QuizCreation is in its own
immersive page when the user is selecting resources, so the
BottomNavigationBar is not present. However, in Lessons, the base Lesson
page is not an immersive page, so the app-mode-specific menu is still
visible, causing the side panel's bottom controls to be hidden behind
the menu. Rather than fiddling w/ z-indexes or trying to conditionally
hide that menu, this just puts the Save & Finish button in a visible
place